sweep with intellij, potential NPE's resolved

+++ src/pers/hal42/data/LongRange.java
@@ -1,15 +1,5 @@
 package pers.hal42.data;
 
-/**
- * Title:        $Source: /cvs/src/net/paymate/data/LongRange.java,v $
- * Description:
- * Copyright:    Copyright (c) 200?
- * Company:      PayMate.net
- *
- * @author PayMate.net
- * @version $Revision: 1.2 $
- */
-
 public class LongRange extends ObjectRange<Long> {
 
   public LongRange() {
@@ -21,11 +11,11 @@ public class LongRange extends ObjectRange<Long> {
   }
 
   public long low() {
-    return (long) one;
+    return one;
   }
 
   public long high() {
-    return (Long) two;
+    return two;
   }
 
 }
